+++ doomgeneric/r_defs.h
@@ -270,7 +270,7 @@ typedef struct
     fixed_t        bbox[2][4];
 
     // If NF_SUBSECTOR its a subsector.
-    unsigned short children[2];
+    unsigned int children[2];
 
 } node_t;
